Benchmarks
* Rework benchmarks
* Move all test related stuff to test-assets
* Update BENCHMARKS.md
