+++ init_wasm.sh
@@ -9,14 +9,14 @@ download_url=$(echo "$latest_release" | grep -o 'https://.*\.zip')
 # Step 3: Add debug info
 mkdir -p src
 version=$(echo $download_url | cut -d / -f8)
-echo "WASM artifact version: ${version}" > src/version.txt
+echo "WASM artifact version: ${version}" > test-assets/wasm/version.txt
 echo "Last updated: $(date)" >> src/version.txt
 
 # Step 4: Curl the download URL and save as release.zip
 curl -L "$download_url" -o release.zip
 
 # Step 5: Unzip release.zip into the src/ directory
-unzip -o release.zip -d src/
+unzip -o release.zip -d test-assets/wasm/
 
 # Step 6: Clean up & info
 rm release.zip
